formatted codebase and update project documentation
