+++ jest.config.js
@@ -1,7 +1,7 @@
 /** @type {import('jest').Config} */
 module.exports = {
-    testMatch: [
-      "**/step_definitions/**/*Steps.js"
-    ],
-    verbose: true,
-  };
+  testMatch: [
+    "**/step_definitions/**/*Steps.js"
+  ],
+  verbose: true,
+};
