+++ step_definitions/dataValidationSteps.js
@@ -1,6 +1,6 @@
 const { defineFeature, loadFeature } = require('jest-cucumber');
 const apiClient = require('../utils/apiClient');
-const {validateField, validateDateOrder} = require( './common_steps.js' )
+const { validateField, validateDateOrder } = require('../utils/commonSteps')
 
 const feature = loadFeature('./features/dataValidation.feature');
 
@@ -26,7 +26,7 @@ defineFeature(feature, (test) => {
     });
   });
 
-  test('Scenario 3: Verify title field is not null or empty in episodes', ({given, then}) => {
+  test('Scenario 3: Verify title field is not null or empty in episodes', ({ given, then }) => {
     given(/^I send a GET request to "(.*)"$/, async (endpoint) => {
       response = (await apiClient.get(endpoint));
       data = response.data.schedule.elements
@@ -42,7 +42,7 @@ defineFeature(feature, (test) => {
     });
   });
 
-  test('Scenario 4: Verify only one live episode exists', ({given, then}) => {
+  test('Scenario 4: Verify only one live episode exists', ({ given, then }) => {
     given(/^I send a GET request to "(.*)"$/, async (endpoint) => {
       response = (await apiClient.get(endpoint));
       data = response.data.schedule.elements
@@ -52,7 +52,7 @@ defineFeature(feature, (test) => {
       count = 0
       data.forEach((item) => {
         episodeLiveKey = item.episode.live
-        if( episodeLiveKey == true) {
+        if (episodeLiveKey == true) {
           count += 1
         };
       });
@@ -61,17 +61,17 @@ defineFeature(feature, (test) => {
     })
   });
 
-test('Scenario 5: Verify transmission dates', ({given, then}) => {
-  given(/^I send a GET request to "(.*)"$/, async (endpoint) => {
-    response = (await apiClient.get(endpoint));
-    data = response.data.schedule.elements
-  });
-
-  then(/^the transmission_start date should be before the transmission_end date for all episodes$/, async () => {
-    data.forEach((item) => {
-      validateDateOrder(item.transmission_start, item.transmission_end)
+  test('Scenario 5: Verify transmission dates', ({ given, then }) => {
+    given(/^I send a GET request to "(.*)"$/, async (endpoint) => {
+      response = (await apiClient.get(endpoint));
+      data = response.data.schedule.elements
     });
-  })
-});
+
+    then(/^the transmission_start date should be before the transmission_end date for all episodes$/, async () => {
+      data.forEach((item) => {
+        validateDateOrder(item.transmission_start, item.transmission_end)
+      });
+    })
+  });
 
 });
